+++ src/wxmac/src/base/wb_rgn.cc
@@ -248,7 +248,7 @@ void wxRegion::SetArc(float x, float y, float w, float h, float start, float end
 
   static double pi;
   if (!pi)
-    pi = 2 * asin(1);
+    pi = 2 * asin((double)1.0);
 
   start = fmod(start, 2*pi);
   end = fmod(end, 2*pi);
